+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 予約にissuedThroughを追加
 - 予約のreservationForを最適化
 - プロダクトのserviceTypeをサービスアウトプットに連携
+- 予約のprogramMembershipUsed型をPermitに変更
 
 ### Deprecated